requirements.txt
pillow~=10.2.0
torchvision==0.15.2
numpy~=1.26.4
tqdm
ftfy
regex
git+https://github.com/openai/CLIP.git
scikit-learn
gdown
torchcp==0.1.3
pandas
openpyxl
transformers==4.50.0